Restrucrured files
 * Moved notebooks in examples
 * Added .gitignore rows
 * Removed useless stuff
