+++ conda.md
@@ -1,5 +1,2 @@
-# Explainable AI
+You must run the repository directly from a local copy of the source code in order to use the repository (without building the package). Installing all required Anaconda and PyPI dependencies is required for this.
 
-## Introduction
-
-This repository contains examples and best practices for building explainable AI systems, provided as Jupyter notebooks.
