--- model_algorithms/README.md
@@ -1,9 +1,0 @@
-# Pretrained models
-
-## Models
-
-##### VGG19 model
-
-A convolutional neural network with 19 layers is called VGG-19. The ImageNet database contains a pretrained version of the network that has been trained on more than a million pictures (check [ImageNet](https://image-net.org/)). The pretrained network can categorize photos into 1000 different item categories, including several animals, a keyboard, a mouse, and a pencil. The network has therefore acquired rich feature representations for a variety of pictures. The network accepts images with a resolution of 224 by 224.
-
-
