--- xai_algorithms/README.md
@@ -1,5 +1,0 @@
-# Explainable AI
-
-## Introduction
-
-
